Update README.md
Healthcare-using-python
To assist you in processing or understanding how data analysis can be carried out, we will cover a range of topics in this repository, such as data cleaning, data modification, data visualisation, and some practical exercises. In this section, we will discuss 26 problem statements and offer the datasets required to solve the issues. If we accomplish this hands-on, the performance would result in an intermediate learning level, making us suitable to start a career as data analysts and data analytics